+++ Src/Seeta/DetectorSeeta.cpp
@@ -14,8 +14,9 @@ QVector<QRect> CDetectorSeeta::Detect(const QImage &image)
     QImage img = image;
     if(img.format() != QImage::Format_RGB888)
     {
-        img = img.convertToFormat(QImage::Format_RGB888).rgbSwapped();
+        img = img.convertToFormat(QImage::Format_RGB888);
     }
+    img = img.rgbSwapped();
     
     SeetaImageData data;
     data.width = img.width();