+++ Src/Seeta/RecognizerSeeta.cpp
@@ -59,8 +59,9 @@ qint64 CRecognizerSeeta::Register(const QImage &image,
     QImage img = image;
     if(img.format() != QImage::Format_RGB888)
     {
-        img = img.convertToFormat(QImage::Format_RGB888).rgbSwapped();
+        img = img.convertToFormat(QImage::Format_RGB888);
     }
+    img = img.rgbSwapped();
 
     SeetaImageData data;
     data.width = img.width();
@@ -99,8 +100,9 @@ qint64 CRecognizerSeeta::Query(/*[in]*/ const QImage &image,
     QImage img = image;
     if(img.format() != QImage::Format_RGB888)
     {
-        img = img.convertToFormat(QImage::Format_RGB888).rgbSwapped();
+        img = img.convertToFormat(QImage::Format_RGB888);
     }
+    img = img.rgbSwapped();
     
     SeetaImageData data;
     data.width = img.width();