+++ Src/database/DataRegister.cpp
@@ -3,6 +3,21 @@
 CDataRegister::CDataRegister(QObject *parent) : QObject(parent)
 {}
 
+CDataRegister::CDataRegister(const CDataRegister &other)
+{
+    m_nIdx = other.m_nIdx;
+    m_nNo = other.m_nNo;
+    m_szName = other.m_szName;
+}
+
+CDataRegister& CDataRegister::operator=(const CDataRegister &other)
+{
+    m_nIdx = other.m_nIdx;
+    m_nNo = other.m_nNo;
+    m_szName = other.m_szName;
+    return *this;
+}
+
 qint64 CDataRegister::getIdx()
 {
     return m_nIdx;