+++ ChangeLog.md
@@ -1,7 +1,8 @@
 - v0.0.3
   + Add database
   + Add register from image
-  
+  + Add recognizer from image
+
 - v0.0.2
   + Register and recognizer face from video file
   + Add format convert